tq: move object.go into transfer.go
